User groups introduced
Each user is a member of some group specifying the time he has
for the contest. This can be useful for off-site participants
taking part at some different time than the others as well as
for test users which should be able to submit even before and
after the contest.