Added logging explanation to two exceptions.
First is "wrong connection string", second is "service not specified
in conf". First still has several occurrence - we fixed only the case
in which the service accepts a contest id.

I introduced a new exception, ConfigError, which is supposed to be
used for critical configuration problems, and is catched only in the
scripts' main.