Fix SmartMappedCollection
Fixes a bug introduced by fa5e98cda752df039abc38cb9fe783e71d6eb0f5 which
caused collections loaded from the DB to be empty (even if the tests
passed). The cause was setting the parent of the child to None inside
__setitem__. The decision to do it came from a wrong understanding of
the dynamics of SQLAlchemy and of the situations that method could be
called in. The new code (that doesn't assume anything about the state
the child is in) should be more correct.

I couldn't find a way to trigger the bug in a testing environment to
make sure to catch it in case it should pop up again.