Make ResourceService work with psutil-3.0
In psutil-3.0 they removed for good some functions that have been
deprecated for a while. Only one affected us, for which there's a handy
substitute, that always existed. (In fact, get_process_list was defined
as list(process_iter()).) Moreover we restored the old minimal required
version of psutil, and updated the highest supported one (because 3.0.1
is working on my system).