Avoid a leak in SmartMappedCollection
It looks like we misunderstood what a dispose event means and thus when
it's fired. In most cases, in fact, it wasn't fired at all and thus the
event listener was never detached. This meant that SQLAlchemy kept a
reference to used and useless SmartMappedCollection instances,
preventing them from being garbage-collected and destroyed. This caused
a leak.

Moreover, as the number of registered event listeners grew linearly in
time, a very quick performance degradation was experienced.