Improve performance of SmartMappedCollection by storing inverse dict
Operations in on_key_update took linear time in the size of the
collection, because they had to check if a value belonged to the
dictionary and what its key was, and did that naively by iteration. Due
to the uniqueness property of values, the dictionary is inversible and
by effectively storing its inverse (indexed by id) we can sensibly speed
up these operations.