Make json hash deterministic
In Rule.py, we use the hash of the json-encoded mission dictionary to
determine if a rule has already been run. The order of the keys in a
dictionary is non-deterministic, so we might do unnecessary reruns. This
commit makes json sort the keys when encoding, so the same mission is
guaranteed to always produce the same hash.