Improve restricted feedback in SubtaskGroup score type
Unless the feedback level is "full", we restrict the information shown to
contestants as follows:

In each group, we show only the first test case with minimum score and we hide
used time and memory. If all test cases in a group get full score, we instead
only show a message "All correct".

This restriction doesn't apply to subtasks that don't contribute to the private
(=official) score. Hence, we give full information about all sample test cases
and about all test cases in detailed feedback subtasks.