ConstraintParser: Implement new requirement for constraints in subscopes
Previously, it was allowed to have a constraint in a subscope (e.g., N<=10000) that was less
strict than the one in its superscope (e.g., N<=100). In that case, only the constraint
of the subscope would be checked, superseding the stricter constraint of its superscope.

Now, an Exception is thrown when a subscope has an explicit constraint that is less strict than one
in its superscope, as this is considered unintentional and there would be no canonical
behaviour for that case, so any specification could be misleading.