Rewrote constraint parsing and added more flexible syntax
You can now specify the TeX code for a given variable or bound by adding the desired code in parentheses directly after, e.g.
constraint(r"sumli(\ell_1+\cdots+\ell_k): [1(\frac{1}{2}+\frac{1}{2}]")
will create the constraint the variable sumli has to be 1, but this condition will be typeset as
$\ell_1+\cdots+\ell_k=\frac{1}{2}+\frac{1}{2}$
in the statement.

Recall, that you already had the possibility to use TeX code for variable names and to some extent in variable bounds. The possibilities for the latter case have been extended independently. So this new feature should be rather seen as providing a shorthand (e.g. for a variable needing some elaborate TeX code to typeset), as a means to reduce dependencies (you do not want to change many constraints and your checker, if you decide to typeset the same variable differently!), or for complicated TeX code in the variable bounds.

In addition you can now enclose variable names or bounds in quotation marks in case they contain any special characters like ',', ':', '[', '(' etc. For example,
constraint('d_{i,j}:[,30]')
is still not valid (it would be interpreted as two different variables, resulting in invalid TeX code), but you can now write
constraint('"d_{i,j}":[,30]')
for this. You can use the same syntax for the specified typesettings, e.g.
constraint(r'proddiff("\prod_{i,j}(x_i-x_j)"): [,1000]')
(this would not have been valid without the quotation marks because of the parentheses in TeX code).