Require Captcha for registration
Implement CaptchaHandler. Include it in registration form.
A registrant has to enter his desired username first to be presented with a captcha. The captcha comes with a cookie that is only valid for a registration with that specific username (and the code that is shown in the captcha picture and has to be correctly entered).