only allow lists as value of primary_statements Some internal functions (e.g. User.__init__) broke when using a dictionary as value of primary_statements. Several other functions included cumbersome case distinctions in order to handle the different possible types of primary_statements.
The original use case for dictionaries was a contest where not all tasks are translated, which is now instead handled by simply skipping missing translations.

Now that primary_statements is a simple list, we can also easily deduce for each user which language versions of the overview sheet he will be interested in.