Don't add the same case multiple times to a test case group
We use subsume_subtask to specify containment relationships
between subtasks. If the corresponding containment graph contains
x paths from subtask A to subtask B, then each test case from B
was added x times to A. That's pointless.

We now don't add a test case to a group if it has been added
before. (Otherwise, the order of test cases is preserved.)