Make LaTeX Sandbox use a persistent LuaLaTeX cache
A cache is copied to the LuaLaTeX sandbox on every run, and then copied
back after the run.

There is a separate sandbox for each 'party', where party is the end of
the name of the LaTeX file that is compiled, e.g. statement-[party].tex.
This will usually be a country (or language) identifier, so the caches
of translators from different countries are kept separate.