GerMake: Don't remake testcases on numbering change
When making a testcase, its number (codename) was used to identify it
and check whether the executable had to actually be called again. So
when the numbering changed, e.g. because a testcase was added or
removed, all testcases that followed had to be re-made even though they
were still present under a different name.

We now identify and store testcases by a hash internally. This means
only this hash instead of the previous testcase number should become
part of the rules missions, which therefore won't change on a change of
numbering.

The hash is calculated as the hash of the rule mission for generating
the testcase input file, reduced by the path of the output (i.e., the
testcase input file).

Numbers are still used like before as the testcases' codenames, and
links from these numbers to the directories actually containing the
testcases are created in the build directory.

To prevent obsolete testcases from staying in the build directory
indefinitely, all testcases that weren't "made" this time are removed
at the end of each run of GerMake (garbage collection).