GerMake(Task): Don't re-evaluate test submissions on testcase numbering change
For local test submissions in GerMake(Task), the testcase codename is omitted
from the evaluation jobs so it doesn't become part of the mission used for
determining if the evaluation job actually has to be run again.

Thus, local test submissions will no longer be re-evaluated on a testcase when
only the testcase's number has changed.

Thanks for the feedback, @fagu!