Prune LaTeX output
Remove everything from the LaTeX output that is not considered to require
attention. Introduce a -vl (--verbose-latex) flag for GerMake and GerMakeTask
to let you see the (almost) complete LaTeX output. The verbose output is also
shown when LaTeX has output an error.

This should make the output of GerMake and GerMakeTask more readable.