Fix updater 29 when unit_test_score_details is not in the dump
The problem is likely that when we introduced unit_test_score_details
we didn't add an updater adding the field with its default value, so
the proper fix would be to create that updater. But this seems close
enough, as it is different only if somebody were to update a dump to a
version of years ago.