Fix submission test
In add_contest, the main_group_id has to be sent when editing the
contest right after creating it to populate it with the desired values.
This was missing, which led to the contest handler quitting instead of
setting the attributes. Because of this, languages=list(ALL_LANGUAGES)
was not set, and no submission languages besides the default ones were
allowed.